Update language.js
Hello Jean! 
Since a few days, I got an error on Chrome & Safari with poggers. 
![](https://puu.sh/EBxVV/51709c8cea.png)

It happens on Safari & Chrome, only on the embed of poggers in MadamNazar.io, and it block the rendering of the map. 

I've found a fix for it, I was wondering if it was possible to add this line: 
```
  if (lang !== undefined) {
    // code ... 
  }
```

It will make sure the `lang` is not `undefined`, and so only fetch the JSON when `lang` is available. 

I've tested locally, it works like a charm. If you could update it, it would be amazing. 


Thanks in advance,
Lucas.